draw ping as integer of milliseconds

+++ game.cpp
@@ -119,7 +119,7 @@ void Game::render(float delta){
 
 	world->drawElements(window);
 
-	latencyDisplayText.setString("Latency: " + std::to_string(network->latency*1000) + "ms\n"
+	latencyDisplayText.setString("Latency: " + std::to_string((int)(network->latency*1000)) + "ms\n"
 		+ "Download: " + std::to_string(network->DISP_TX/1000.0) + "kB/s\n"
 		+ "Upload:   " + std::to_string(network->DISP_RX/1000.0) + "kB/s\n");
 	latencyDisplayText.setPosition(
